pyproject.toml
[project]
name = "thelasthours"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas",
    "spacey",
    "matplotlib",
    "selenium",
    "networkx",
    "numpy",
    "webdriver_manager.chrome"
]
